feat: save Purchase and SeatPurchase
