+++ src/main/java/com/distinction/barbenheimer/controller/PaymentController.java
@@ -1,20 +1,16 @@
 package com.distinction.barbenheimer.controller;
 
 import com.distinction.barbenheimer.DTO.PaymentDTO;
-import com.distinction.barbenheimer.service.PaymentService;
-
-import com.stripe.Stripe;
+import com.distinction.barbenheimer.service.SeatPurchaseService;
 import com.stripe.exception.StripeException;
-import com.stripe.model.checkout.Session;
-import com.stripe.param.checkout.SessionCreateParams;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
@@ -23,28 +19,19 @@ import java.util.List;
 @Slf4j
 public class PaymentController {
 
-    private final PaymentService paymentService;
+    private final SeatPurchaseService seatPurchaseService;
 
     @Autowired
-    public PaymentController(PaymentService paymentService){
-        this.paymentService = paymentService;
+    public PaymentController(SeatPurchaseService seatPurchaseService){
+        this.seatPurchaseService = seatPurchaseService;
     }
-
-    //Things this needs: Customer Detail and Seat Selected
-    // upon landing on this payment page there will be a timeout after 10 mins.
-//    @GetMapping
-//    public ResponseEntity<?> awaitPurchase(HttpSession httpSession){
-//        httpSession.setAttribute("customerToken", seatPurchaseService.createCustomerIdentifyingToken());
-//        httpSession.setMaxInactiveInterval(600);
-//        return ResponseEntity.ok(httpSession.getAttribute("customerToken"));
-//    }
 
     @PostMapping("/create-checkout-session")
-    public String checkout (@RequestBody List<PaymentDTO> paymentDTOList, HttpServletResponse response) throws StripeException, IOException {
-        return paymentService.checkout(paymentDTOList, response);
+    public String checkout (@RequestBody PaymentDTO paymentDTO, HttpServletResponse response, HttpSession httpSession) throws StripeException, IOException {
+        httpSession.setAttribute("customerToken", seatPurchaseService.createCustomerIdentifyingToken());
+//        httpSession.setAttribute("seatSelectDTO", seatSelectDTO);
+        return seatPurchaseService.checkout(paymentDTO, response, httpSession);
     }
-
-
 
 //    @GetMapping
 //    public ResponseEntity<?> purchaseSeats(@RequestBody SeatSelectDTO seatSelectDTO, HttpSession httpSession){
@@ -59,7 +46,12 @@ public class PaymentController {
 //        return ResponseEntity.ok(seatSelectDTO.toString());
 //    }
 
-
+//    @GetMapping
+//    public ResponseEntity<?> awaitPurchase(HttpSession httpSession){
+//        httpSession.setAttribute("customerToken", seatPurchaseService.createCustomerIdentifyingToken());
+//        httpSession.setMaxInactiveInterval(600);
+//        return ResponseEntity.ok(httpSession.getAttribute("customerToken"));
+//    }
 
 
 }
